Moves internal collection related stuff into brave.internal.collect
It is easy to lose track of our goodies. This moves them before
re-introducing a map base type for Extra.
