+++ brave/src/main/java/brave/internal/collect/Lists.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package brave.internal;
+package brave.internal.collect;
 
 import java.util.ArrayList;
 import java.util.Arrays;
