+++ brave/src/main/java/brave/internal/collect/WeakConcurrentMap.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package brave.internal.weaklockfree;
+package brave.internal.collect;
 
 import brave.internal.Nullable;
 import java.lang.ref.Reference;
