+++ brave/src/test/java/brave/internal/collect/ListsTest.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package brave.internal;
+package brave.internal.collect;
 
 import com.google.common.collect.ImmutableList;
 import java.util.ArrayList;
