+++ brave/src/test/java/brave/internal/collect/UnsafeArrayMapTest.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package brave.internal.baggage;
+package brave.internal.collect;
 
 import java.util.Collection;
 import java.util.Collections;
